requirements.txt
nextcord
google-genai
python-dotenv
aiohttp
aiofiles
duckduckgo-search
beautifulsoup4
RestrictedPython
aiosqlite
onami
ddgs
pytz
requests
PyMuPDF
python-docx
math2docx
